resolving comment and conflicts
